% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_annual_mean}
\alias{cnr_annual_mean}
\title{Compute means for layers in rast in a given year}
\usage{
cnr_annual_mean(r, y)
}
\arguments{
\item{r}{rast}

\item{y}{year}
}
\value{
rast with mean
}
\description{
Compute means for layers in rast in a given year
}
