% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_crs}
\alias{cnr_crs}
\title{Default CRS for specific nc files}
\usage{
cnr_crs(var, fn)
}
\arguments{
\item{var}{the variable}

\item{fn}{filename pattern}
}
\description{
Default CRS for specific nc files
}
