% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_dir}
\alias{cnr_dir}
\title{File path for scratch dir with nc files}
\usage{
cnr_dir(subpath, scratchdir = "/media/markus/Stash/data/out")
}
\arguments{
\item{subpath}{subpath}

\item{scratchdir}{default location of nc files}
}
\description{
File path for scratch dir with nc files
}
