% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_dir_meta}
\alias{cnr_dir_meta}
\title{Data frame with nc files listing with some metadata added,
extracted from the filenames}
\usage{
cnr_dir_meta(path, pat = "*.nc$")
}
\arguments{
\item{path}{path}

\item{pat}{pattern}
}
\description{
Data frame with nc files listing with some metadata added,
extracted from the filenames
}
