% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_ext_df}
\alias{cnr_ext_df}
\title{Extract coordinate data from data frame}
\usage{
cnr_ext_df(data, x, y)
}
\arguments{
\item{data}{data}

\item{x}{x coord}

\item{y}{y coord}
}
\description{
Extract coordinate data from data frame
}
