% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_ext_range_df}
\alias{cnr_ext_range_df}
\title{Wrapper function to extract data within specific bounds}
\usage{
cnr_ext_range_df(data, xmin, xmax, ymin, ymax)
}
\arguments{
\item{data}{data}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}
}
\description{
Wrapper function to extract data within specific bounds
}
