% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_extract_range}
\alias{cnr_extract_range}
\title{Query a parquet file for data within (inclusive) a certain extent}
\usage{
cnr_extract_range(pf, ymin, ymax, xmin, xmax, pd = cnr_dir("/tdc"))
}
\arguments{
\item{pf}{parquet file}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{pd}{parquet file directory}
}
\description{
Query a parquet file for data within (inclusive) a certain extent
}
