% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_extract_xy}
\alias{cnr_extract_xy}
\title{Wrapper function to extract data at a specific coordinate}
\usage{
cnr_extract_xy(pf, y, x)
}
\arguments{
\item{pf}{parquet file}

\item{y}{y}

\item{x}{x}
}
\description{
Wrapper function to extract data at a specific coordinate
}
