% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_filename}
\alias{cnr_filename}
\title{Filename(s) on the form variable_year_month
(useful for naming several output layers in a netcdf file)}
\usage{
cnr_filename(md)
}
\arguments{
\item{md}{metadata from cnr_ncmeta}
}
\description{
Filename(s) on the form variable_year_month
(useful for naming several output layers in a netcdf file)
}
