% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_filenames}
\alias{cnr_filenames}
\title{Filenames from rast layers}
\usage{
cnr_filenames(r, extension)
}
\arguments{
\item{r}{rast}

\item{extension}{file extension}
}
\description{
Filenames from rast layers
}
