% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_flatten_old}
\alias{cnr_flatten_old}
\title{Flatten layers in a raster across cells into a single layer
where each cell takes the first value which is not NaN (ie is numerical)}
\usage{
cnr_flatten_old(r)
}
\arguments{
\item{r}{rast}
}
\value{
flattened raster
}
\description{
Flatten layers in a raster across cells into a single layer
where each cell takes the first value which is not NaN (ie is numerical)
}
