% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_flatten_save}
\alias{cnr_flatten_save}
\title{Given ncfiles, flatten layers and save as individual nc files}
\usage{
cnr_flatten_save(ncfiles, reproject_from, prefix, outdir = cnr_dir("orig"))
}
\arguments{
\item{ncfiles}{vector of ncdf files}

\item{reproject_from}{crs to project from}

\item{prefix}{variable name for prefixing file output}

\item{outdir}{location of resulting files}
}
\value{
flattened raster
}
\description{
Given ncfiles, flatten layers and save as individual nc files
}
