% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_mismatch_tdc}
\alias{cnr_mismatch_tdc}
\title{Find locations which do not match with the provided template
(note that there may possibly be mismatches due to approximations/rounding)}
\usage{
cnr_mismatch_tdc(r, type = c("rast", "tibble"))
}
\arguments{
\item{r}{rast}

\item{type}{one of rast or tibble}
}
\description{
Find locations which do not match with the provided template
(note that there may possibly be mismatches due to approximations/rounding)
}
