% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_monthlies}
\alias{cnr_monthlies}
\title{Index a set of raster layers by year}
\usage{
cnr_monthlies(r, y)
}
\arguments{
\item{r}{rast}

\item{y}{year}
}
\description{
Index a set of raster layers by year
}
