% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_ncmeta}
\alias{cnr_ncmeta}
\title{Metadata from nc file}
\usage{
cnr_ncmeta(fp, var, fn)
}
\arguments{
\item{fp}{filepath}

\item{var}{variable}

\item{fn}{filename}
}
\description{
Inside a netcdf file there is metadata embedded, which can
be parsed with the \code{ncmeta} R package.
}
