% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_outer}
\alias{cnr_outer}
\title{A table of variable / year combinations}
\usage{
cnr_outer(vars, years, nm = c("var", "year", "fn"))
}
\arguments{
\item{vars}{variable names}

\item{years}{vector of years}

\item{nm}{header names}
}
\description{
A table of variable / year combinations
}
