% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_process2}
\alias{cnr_process2}
\title{Process a list of raster with slots for variables}
\usage{
cnr_process2(r_all, myvar)
}
\arguments{
\item{r_all}{rast}

\item{myvar}{variable to subset on}
}
\description{
Process a list of raster with slots for variables
}
