% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_raster}
\alias{cnr_raster}
\title{Retrieve raster from a nc file (which may have several layers) where
translation of values is optional because}
\usage{
cnr_raster(f, translate = FALSE, rename_layers = TRUE)
}
\arguments{
\item{f}{file}

\item{translate}{boolean}

\item{rename_layers}{boolean}
}
\description{
\enumerate{
\item it takes time and could be delayed until after resampling
\item results from terra seems to lose some original metadata in the translation
}
}
