% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_reproject}
\alias{cnr_reproject}
\title{Reproject a layer
where each cell takes the first value which is not NaN (ie is numerical)}
\usage{
cnr_reproject(r, crs_from)
}
\arguments{
\item{r}{rast}

\item{crs_from}{valid EPSG crs string}
}
\value{
flattened raster
}
\description{
Reproject a layer
where each cell takes the first value which is not NaN (ie is numerical)
}
