% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_resample}
\alias{cnr_resample}
\title{Resample raster to tenth degree resolution cells
with optional transformation of values
(using offset and scale from metadata)}
\usage{
cnr_resample(f, nth_degrees = 10, transform = TRUE, type = c("file", "rast"))
}
\arguments{
\item{f}{nc file path or rast}

\item{nth_degrees}{nth_degrees}

\item{transform}{boolean}

\item{type}{one of "file" or "rast"}
}
\description{
Resample raster to tenth degree resolution cells
with optional transformation of values
(using offset and scale from metadata)
}
