% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_csv}
\alias{cnr_save_csv}
\title{Output a terra layer to a compressed CSV file
(using .gz compression which is twice as fast compared to .zip)}
\usage{
cnr_save_csv(r, outfile, compress = TRUE)
}
\arguments{
\item{r}{rast}

\item{outfile}{outfile}

\item{compress}{boolean}
}
\description{
Output a terra layer to a compressed CSV file
(using .gz compression which is twice as fast compared to .zip)
}
