% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_csvs}
\alias{cnr_save_csvs}
\title{Write .csv files for variables across time}
\usage{
cnr_save_csvs(f, path = cnr_dir())
}
\arguments{
\item{f}{file}

\item{path}{directory for output files}
}
\description{
Write .csv files for variables across time
}
