% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_csvz}
\alias{cnr_save_csvz}
\title{Write .csv files for variables across time}
\usage{
cnr_save_csvz(r, path = cnr_dir())
}
\arguments{
\item{r}{rast}

\item{path}{output directory}
}
\description{
Write .csv files for variables across time
}
