% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_html}
\alias{cnr_save_html}
\title{Save outputs from raster to html file}
\usage{
cnr_save_html(r, outfile)
}
\arguments{
\item{r}{rast}

\item{outfile}{outfile}
}
\description{
Save outputs from raster to html file
}
