% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_htmls}
\alias{cnr_save_htmls}
\title{Save raster layers in nc file to a directory as html files}
\usage{
cnr_save_htmls(f, path = cnr_dir(), resample = TRUE)
}
\arguments{
\item{f}{file}

\item{path}{output dir}

\item{resample}{boolean}
}
\description{
Save raster layers in nc file to a directory as html files
}
