% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_mapview}
\alias{cnr_save_mapview}
\title{Save mapview file for a raster to a specified file}
\usage{
cnr_save_mapview(r, outfile)
}
\arguments{
\item{r}{rast}

\item{outfile}{outfile}
}
\description{
Save mapview file for a raster to a specified file
}
