% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_mapviews}
\alias{cnr_save_mapviews}
\title{Save raster layers as mapviews in a specified dir}
\usage{
cnr_save_mapviews(r, outdir = "~/data/out/orig")
}
\arguments{
\item{r}{rast}

\item{outdir}{output directory}
}
\description{
Save raster layers as mapviews in a specified dir
}
