% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_monthlies}
\alias{cnr_save_monthlies}
\title{Save yearly tables with monthly columns in an output directory as .csv.gz}
\usage{
cnr_save_monthlies(r_unified, var, y, outdir, compress = TRUE)
}
\arguments{
\item{r_unified}{rast}

\item{var}{variable}

\item{y}{year}

\item{outdir}{outdir}

\item{compress}{boolean}
}
\description{
Save yearly tables with monthly columns in an output directory as .csv.gz
}
