% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_tiff}
\alias{cnr_save_tiff}
\title{Output a terra layer to TIFF (using terra)}
\usage{
cnr_save_tiff(r, outfile, ...)
}
\arguments{
\item{r}{rast}

\item{outfile}{outfile}

\item{...}{args to send to terra::writeRaster}
}
\description{
Output a terra layer to TIFF (using terra)
}
