% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_tiffs}
\alias{cnr_save_tiffs}
\title{Output all layers in a nc file to TIFF (using terra)}
\usage{
cnr_save_tiffs(f, path = cnr_dir())
}
\arguments{
\item{f}{file}

\item{path}{output directory}
}
\description{
Output all layers in a nc file to TIFF (using terra)
}
