% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_tiffz}
\alias{cnr_save_tiffz}
\title{Output all layers in a nc file to TIFF (using terra)}
\usage{
cnr_save_tiffz(r, path = cnr_dir())
}
\arguments{
\item{r}{rast}

\item{path}{output directory}
}
\description{
Output all layers in a nc file to TIFF (using terra)
}
