% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_save_ym}
\alias{cnr_save_ym}
\title{Write a set of raster layers by year with monthly columns as CSV}
\usage{
cnr_save_ym(r, y, varname, outdir, compress = FALSE)
}
\arguments{
\item{r}{rast}

\item{y}{year}

\item{varname}{variable name}

\item{outdir}{output directory}

\item{compress}{boolean}
}
\description{
Write a set of raster layers by year with monthly columns as CSV
}
