% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_terrameta}
\alias{cnr_terrameta}
\title{In addition to the metadata embedded in nc file names and to what \code{ncmeta}
can provide, there is also metadata available through inspection of
\code{terra} objects.}
\usage{
cnr_terrameta(f)
}
\arguments{
\item{f}{filepath to nc file}
}
\description{
Metadata from \code{terra} given a file
}
