% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_transform}
\alias{cnr_transform}
\title{Transform given scale factor and offset}
\usage{
cnr_transform(r, var)
}
\arguments{
\item{r}{rast}

\item{var}{variable}
}
\description{
Transform given scale factor and offset
}
