% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_transform_params}
\alias{cnr_transform_params}
\title{The nc metadata for scale factor and offset for a given variable}
\usage{
cnr_transform_params(myvar)
}
\arguments{
\item{myvar}{variable name}
}
\description{
The nc metadata for scale factor and offset for a given variable
}
