% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_unified}
\alias{cnr_unified}
\title{Unify several nc files into a raster with data
for a specific variable (multi-layer/"brick")}
\usage{
cnr_unified(filez, varname)
}
\arguments{
\item{filez}{vector of file paths}

\item{varname}{varname}
}
\description{
Unify several nc files into a raster with data
for a specific variable (multi-layer/"brick")
}
