% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{cnr_view_parquet}
\alias{cnr_view_parquet}
\title{View first few rows in a parquet file}
\usage{
cnr_view_parquet(fn, n_row = 10)
}
\arguments{
\item{fn}{filename}

\item{n_row}{nr of rows to preview}
}
\description{
View first few rows in a parquet file
}
