% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnr.R
\name{nc_file}
\alias{nc_file}
\title{Docs for terra::rast() says to prepend format
so netCDF files "do not open with the HDF5 driver".}
\usage{
nc_file(file)
}
\arguments{
\item{file}{file}
}
\description{
Docs for terra::rast() says to prepend format
so netCDF files "do not open with the HDF5 driver".
}
